// at-backend/src/common/enums/conflict-reason-code.enum.ts
export enum ConflictReasonCode {
  SHIFTED_BY_WINDOW_CONFLICT = 'SHIFTED_BY_WINDOW_CONFLICT',
  INACTIVATED_BY_MANDATORY_RULE = 'INACTIVATED_BY_MANDATORY_RULE',
  MANUAL_REQUIRED_PERSISTENT_CONFLICT = 'MANUAL_REQUIRED_PERSISTENT_CONFLICT',
  MANUAL_REQUIRED_NON_MOVABLE_COLLISION = 'MANUAL_REQUIRED_NON_MOVABLE_COLLISION',
  MANUAL_REQUIRED_ITERATION_LIMIT = 'MANUAL_REQUIRED_ITERATION_LIMIT',
  MANUAL_REQUIRED_MISSING_ROUTINE_ANCHOR = 'MANUAL_REQUIRED_MISSING_ROUTINE_ANCHOR',
  SHIFTED_BY_OPHTHALMIC_INTERVAL = 'SHIFTED_BY_OPHTHALMIC_INTERVAL',
  MANUAL_REQUIRED_OPHTHALMIC_INTERVAL = 'MANUAL_REQUIRED_OPHTHALMIC_INTERVAL',
}